+++ packages/core/lib/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.test.ts
