+++ packages/core/lib/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.ts
