+++ packages/core/lib/getDisposalTextFromResult/exclusionOrderDisposalText.ts
