+++ packages/core/lib/getDisposalTextFromResult/exclusionRequirementsDisposalText.test.ts
