+++ packages/core/lib/getDisposalTextFromResult/exclusionRequirementsDisposalText.ts
