+++ packages/core/lib/getDisposalTextFromResult/index.ts
