+++ packages/core/lib/getDisposalTextFromResult/isFailedToAppearWarrantIssued.test.ts
