+++ packages/core/lib/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
