+++ packages/core/lib/getDisposalTextFromResult/licencedPremisesExclusionOrderDisposalText.ts
