+++ packages/core/lib/getDisposalTextFromResult/penaltyPointsDisposalText.ts
@@ -1,4 +1,4 @@
-import type { Result } from "../../../types/AnnotatedHearingOutcome"
+import type { Result } from "../../types/AnnotatedHearingOutcome"
 
 const penaltyPointsDisposalText = (result: Result): string => {
   const numberSpecifiedInResultValue = result.NumberSpecifiedInResult?.find((num) => num.Number)?.Number
