+++ packages/core/lib/getDisposalTextFromResult/untilFurtherOrderDisposalText.ts
